#19 opening 컴포넌트 분리 및 공통화 작업 중.
